% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_tukey_graficas_qRT_PCR.R
\name{anova_tukey_graficas_qRT_PCR}
\alias{anova_tukey_graficas_qRT_PCR}
\title{Hace Analisis Estadistico de ANOVA y TUKEY. Hace Grafica (de barras) de Tratamiento/Condicion vs Control/Normalizador con Desviacion Estandar Conjunta de Calidad de Publicacion}
\usage{
anova_tukey_graficas_qRT_PCR(
  ruta,
  tratamiento_condicion,
  target,
  normalizador,
  tratamiento
)
}
\arguments{
\item{ruta}{Ruta de carpeta origen en winsys}

\item{tratamiento_condicion}{Tratamiento y condicion (por ejemplo: Tratamiento: NaCl. Condicion: 3h)}

\item{target}{Gen problema (por ejemplo cornichon CN1)}

\item{normalizador}{Gen normalizador (por ejemplo tubulina, Tb o factor de elongacion, Ef)}

\item{tratamiento}{Tratamiento (por ejemplo estres abiotico, NaCl)}
}
\value{
Graficos y resumen de resultados de anova
}
\description{
Trabaja sobre objetos de R previamente creados.
Hace analisis estadistico de ANOVA, guarda resultados en winsys
Hace analisis estadistico de TUKEY, guarda plot en winsys
Hace GRAFICO DE BARRAS de tratamiento vs control, con desviacion estandar conjunta, guarda resultados en winsys
}
