% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinar_promediar_qRT_PCR.R
\name{combinar_promediar_qRT_PCR}
\alias{combinar_promediar_qRT_PCR}
\title{Combina datos de DDCT y de Desviacion Estandar para Analisis de qRT PCR.}
\usage{
combinar_promediar_qRT_PCR(ruta, target, normalizador)
}
\arguments{
\item{ruta}{Ruta de carpeta origen en winsys}

\item{target}{Gen problema (por ejemplo cornichon CN1)}

\item{normalizador}{Gen normalizador (por ejemplo tubulina, Tb, o factor de elongacion, Ef)}
}
\value{
Datos combinados por las diferentes corridas, listos para graficacion y analisis
}
\description{
Trabaja sobre objetos de R previamente creados.
Inicia analisis y transformacion de datos por corrida
Combina datos de desviacion estandar conjunta (control y tratamiento/condicion)
Hace grafico de barras de tratamiento vs control, con desviacion estandar conjunta, guarda resultados en winsys
}
