% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinar_tratamientos_condiciones_qRT_PCR.R
\name{combinar_tratamientos_condiciones_qRT_PCR}
\alias{combinar_tratamientos_condiciones_qRT_PCR}
\title{Combina los Diferentes Tratamientos y Condiciones en un Analisis Global de qRT PCR. Genera sus Propias Tablas, Archivos de Resultados y Graficas de ANOVA, Expresion Relativa y Tukey}
\usage{
combinar_tratamientos_condiciones_qRT_PCR(
  ruta_carpeta,
  vector_subcarpetas,
  tratamiento_condiciones,
  target,
  normalizador
)
}
\arguments{
\item{target}{Gen problema (por ejemplo cornichon CN1)}

\item{normalizador}{Gen normalizador (por ejemplo tubulina, Tb o factor de elongacion, Ef)}

\item{ruta}{Ruta de carpeta origen en winsys}

\item{tratamiento_condicion}{Tratamiento y condicion (por ejemplo: Tratamiento: NaCl. Condicion: 3h)}
}
\value{
Graficos y resumen de resultados de anova
}
\description{
Guarda los objetos de R principales cobinados por tratamiento y condicion, en winsys
Incluyen:
Tablas usadas para graficas
Tablas usadas para analisis estadistico de ANOVA y Tukey
}
