% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniciar_qRT_PCR.R
\name{iniciar_qRT_PCR}
\alias{iniciar_qRT_PCR}
\title{Filtra y Transforma Datos de las Diferentes Corridas. Obtiene Delta Delta CT y Desviaciones Estandar (ds)}
\usage{
iniciar_qRT_PCR(ruta, archivos, tratamiento, target, normalizador)
}
\arguments{
\item{ruta}{Ruta de carpeta origen en winsys}

\item{archivos}{Vector de archivos de las diferentes}

\item{tratamiento}{Tratamiento (por ejemplo estres abiotico, NaCl)}

\item{target}{Gen problema (por ejemplo cornichon CN1)}

\item{normalizador}{Gen normalizador (por ejemplo tubulina, Tb o factor de elongacion, Ef)}
}
\value{
Obtiene y filtra datos de corridas para analisis de qRT PCR (separados por corrida)
}
\description{
Filtra y transforma datos por corrida
}
